Refactor the observer functions into a plugin
